consolidate db.roles module
