+++ mathesar/rpc/databases/privileges.py
@@ -3,9 +3,11 @@ from typing import Literal, TypedDict
 from modernrpc.core import rpc_method, REQUEST_KEY
 from modernrpc.auth.basic import http_basic_auth_login_required
 
-from db.roles.operations.select import list_db_priv
-from db.roles.operations.update import replace_database_privileges_for_roles
-from db.roles.operations.ownership import transfer_database_ownership
+from db.roles import (
+    list_db_priv,
+    replace_database_privileges_for_roles,
+    transfer_database_ownership,
+)
 from mathesar.rpc.databases.base import DatabaseInfo
 from mathesar.rpc.utils import connect
 from mathesar.rpc.exceptions.handlers import handle_rpc_exceptions
